tg/requirements.txt
python-telegram-bot
pydub
dotenv
quart